requirements.txt
django==4.2.*
Pillow==11.0.0
python-dotenv==1.0.1
django-rest-framework